support iframe type syncs
